Promote substreams to next to view.
