+++ src/card-controller/actions/actions/substream-off.ts
@@ -1,6 +1,6 @@
-import { SubstreamViewModifier } from '../../../components-lib/live/substream';
 import { GeneralActionConfig } from '../../../config/schema/actions/custom/general';
 import { CardActionsAPI } from '../../types';
+import { SubstreamViewModifier } from '../../view/modifiers/substream';
 import { AdvancedCameraCardAction } from './base';
 
 export class SubstreamOffAction extends AdvancedCameraCardAction<GeneralActionConfig> {
