+++ src/card-controller/actions/actions/substream-select.ts
@@ -1,6 +1,6 @@
-import { SubstreamViewModifier } from '../../../components-lib/live/substream';
 import { SubstreamSelectActionConfig } from '../../../config/schema/actions/custom/substream-select';
 import { CardActionsAPI } from '../../types';
+import { SubstreamViewModifier } from '../../view/modifiers/substream';
 import { AdvancedCameraCardAction } from './base';
 
 export class SubstreamSelectAction extends AdvancedCameraCardAction<SubstreamSelectActionConfig> {
