+++ src/components/live/carousel.ts
@@ -14,7 +14,6 @@ import { CameraManagerCameraMetadata } from '../../camera-manager/types.js';
 import { CallSession } from '../../card-controller/call/types.js';
 import { MicrophoneState } from '../../card-controller/types.js';
 import { ViewManagerEpoch } from '../../card-controller/view/types.js';
-import { getStreamCameraID } from '../../components-lib/live/substream.js';
 import { MediaActionsController } from '../../components-lib/media-actions-controller.js';
 import { MediaHeightController } from '../../components-lib/media-height-controller.js';
 import { MediaLoadedInfoSinkController } from '../../components-lib/media-loaded-info-sink-controller.js';
@@ -33,6 +32,7 @@ import liveCarouselStyle from '../../scss/live-carousel.scss';
 import { stopEventFromActivatingCardWideActions } from '../../utils/action.js';
 import { CarouselSelected } from '../../utils/embla/carousel-controller.js';
 import { getTextDirection } from '../../utils/text-direction.js';
+import { getStreamCameraID } from '../../view/substream.js';
 import { View } from '../../view/view.js';
 import '../call-controls.js';
 import '../carousel';
