+++ src/utils/ptz.ts
@@ -1,7 +1,7 @@
 import { CameraManager } from '../camera-manager/manager';
-import { getStreamCameraID } from '../components-lib/live/substream';
 import { PTZAction } from '../config/schema/actions/custom/ptz';
 import { PTZCapabilities } from '../types';
+import { getStreamCameraID } from '../view/substream';
 import { getViewTargetID } from '../view/target-id';
 import { View } from '../view/view';
